complete restructuring of the stream module into separate sub modules
