+++ src/Sound/Tidal/Safe/Context.hs
@@ -61,8 +61,9 @@ import Sound.Tidal.ParseBP as C
 import Sound.Tidal.Pattern as C
 import Sound.Tidal.Scales as C
 import Sound.Tidal.Simple as C
-import Sound.Tidal.Stream.Stream
-  (startTidal, superdirtTarget, Target(..))
+import Sound.Tidal.Stream.Target (superdirtTarget)
+import Sound.Tidal.Stream.Types (Target(..))
+import Sound.Tidal.Stream.Main (startTidal)
 -- import Sound.Tidal.Transition as C
 import Sound.Tidal.UI as C
 import Sound.Tidal.Version as C
