+++ tidal.cabal
@@ -41,9 +41,14 @@ library
                        Sound.Tidal.Safe.Boot
                        Sound.Tidal.Show
                        Sound.Tidal.Simple
-                       Sound.Tidal.Stream.Stream
-                       Sound.Tidal.Stream.Types
+                       Sound.Tidal.Stream
                        Sound.Tidal.Stream.Config
+                       Sound.Tidal.Stream.Listen
+                       Sound.Tidal.Stream.Main
+                       Sound.Tidal.Stream.Process
+                       Sound.Tidal.Stream.Types
+                       Sound.Tidal.Stream.Target
+                       Sound.Tidal.Stream.UI
                        Sound.Tidal.Time
                        Sound.Tidal.Transition
                        Sound.Tidal.UI
